pyproject.toml
[project]
name = "citation-retrieval"
version = "0.1.0"
description = "Multi-baseline citation retrieval system with BM25, Dense Retrieval, and LLM-based agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core dependencies
    "pandas>=2.2.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "langchain-anthropic>=0.1.23",
    "langchain-openai>=0.1.25",
    "langchain-together>=0.1.5",
    "pydantic>=2.12.3",
    "pypdf2>=3.0.1",
    "rich>=14.2.0",
    "selenium>=4.38.0",
    "datasets>=4.4.1",
    # BM25 baseline dependencies
    "bm25s>=0.1.0",
    "pystemmer>=2.2.0",
    "matplotlib>=3.8.0",
    # Dense retrieval baseline dependencies
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "peft>=0.7.0",
    # Evaluation framework dependencies
    # (mostly covered above, adding any extras)
    "ipykernel>=7.1.0",
    "langchain-chroma>=0.2.2",
    "langchain>=0.2.17",
    "langchain-community>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["evaluation"]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
